python/requirements.txt
confini>=0.3.6rc3,<0.5.0
crypto-dev-signer>=0.4.14b7,<=0.4.14
chainlib-eth>=0.0.7a4,<=0.1.0
potaahto~=0.0.1a2
